Keep the health smoke gate warning-clean
The first Jetson P0-010 run passed functionally but exposed a format-string warning and truncated health reasons in the generated report. This tightens the smoke gate so package stderr output is rejected and preserves full reason text for later evidence records.

Constraint: P0-010 should not be closed on a build that reports compiler stderr output, even when the smoke verdict is PASS.

Confidence: high

Scope-risk: narrow

Directive: Re-run the P0-010 Jetson smoke after this commit before recording completion evidence.

Tested: powershell verify_p0_010_health_monitor.ps1; bash -n run_p0_010_health_monitor_smoke.sh; powershell verify_p0_010_completion_gate.ps1 on pass fixture; fail fixture rejected; git diff --cached --check

Not-tested: Jetson runtime smoke after the warning-format fix.

+++ ros2_ws/src/edge_reliability_health/src/health_monitor.cpp
@@ -42,8 +42,8 @@ public:
     if (min_expected_count < 0) {
       RCLCPP_WARN(
         get_logger(),
-        "event=parameter_fallback parameter=min_expected_count value=%d fallback=10",
-        min_expected_count);
+        "event=parameter_fallback parameter=min_expected_count value=%ld fallback=10",
+        static_cast<long>(min_expected_count));
       thresholds_.min_expected_count = 10;
     } else {
       thresholds_.min_expected_count = static_cast<uint64_t>(min_expected_count);
